requirements.txt
# Core
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Dynamic crawling
playwright==1.43.0

# AI Layer
google-generativeai==0.5.4

# Data validation
pydantic==2.7.1

# Database
pymongo==4.7.2

# API
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Branding / Image
Pillow==10.3.0
colorthief==0.2.1
tinycss2==1.3.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.0
